Create simple-cards
Card game where cards are split into two decks. The sum of one deck is measured and the product of the other deck is measured. Simple genetic algorithm is employed to work towards the end goal.

Program outputs a console window showing success/failure, number of iterations required, and the arrangement of cards.